# server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend Server for Legacy App",
        "title": "Legacy API",
        "contact": {
            "name": "David Oduneye and Akshay Dupuguntla",
            "email": "oduneye.d@northeastern.edu and dupuguntla@northeastern.edu"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/users/": {
            "get": {
                "description": "Returns all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Gets all users",
                "operationId": "get-all-users",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        }
                    },
                    "404": {
                        "description": "Failed to fetch users",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Creates a user",
                "operationId": "post-user",
                "parameters": [
                    {
                        "description": "User",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Failed to create user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/firebaseid/{firebaseid}": {
            "get": {
                "description": "Returns a user from firebase id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Gets a user from firebase id",
                "operationId": "get-user-from-firebaseid",
                "parameters": [
                    {
                        "type": "string",
                        "description": "FirebaseID",
                        "name": "firebaseid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/username/{username}": {
            "get": {
                "description": "Returns a user from username",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Gets a user from username",
                "operationId": "get-user-from-username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{uid}": {
            "get": {
                "description": "Returns a user from user id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Gets a user from user id",
                "operationId": "get-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Deletes a user",
                "operationId": "delete-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User successfully deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Failed to delete user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Updates a user",
                "operationId": "update-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/docmodels.UserDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Failed to update user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{uid}/persona": {
            "get": {
                "description": "Returns a persona from user id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Gets a user persona from user id",
                "operationId": "get-user-persona",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Persona"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch user persona",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{uid}/profile": {
            "get": {
                "description": "Returns a profile from user id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Gets a user profile from user id",
                "operationId": "get-user-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Profile"
                        }
                    },
                    "404": {
                        "description": "Failed to fetch user profile",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/users/{uid}/progress": {
            "post": {
                "description": "Initializes all tasks and subtasks for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Initializes user progress",
                "operationId": "initialize-user-progress",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UserID",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Failed to initialize user progress",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docmodels.UserDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "firebase_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "persona_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.File": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "object_key": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.Guide": {
            "type": "object",
            "required": [
                "author",
                "author_image_url",
                "full_text",
                "guide_name",
                "sub_title",
                "title"
            ],
            "properties": {
                "author": {
                    "type": "string"
                },
                "author_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "date": {
                    "type": "string"
                },
                "full_text": {
                    "type": "string"
                },
                "guide_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "mins_read": {
                    "type": "integer"
                },
                "sub_title": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.Persona": {
            "type": "object",
            "required": [
                "persona_description",
                "persona_title"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "persona_description": {
                    "type": "string"
                },
                "persona_title": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.Profile": {
            "type": "object",
            "required": [
                "date_of_birth",
                "name",
                "phone_number",
                "user_id"
            ],
            "properties": {
                "completed_onboarding_response": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "date_of_birth": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                },
                "onboarding_response": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.Tag": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.File"
                    }
                },
                "guides": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Guide"
                    }
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Task"
                    }
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
        "models.Task": {
            "type": "object",
            "required": [
                "task_description",
                "task_name"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "personas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Persona"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "task_description": {
                    "type": "string"
                },
                "task_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "required": [
                "email",
                "firebase_id",
                "password",
                "username"
            ],
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "email": {
                    "type": "string"
                },
                "firebase_id": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "password": {
                    "type": "string"
                },
                "persona_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2023-09-20T16:34:50Z"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}